Removed unnecessary cast
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@3445 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/be/ibridge/kettle/core/value/ValueBigNumber.java
@@ -112,7 +112,7 @@ public class ValueBigNumber implements ValueInterface, Cloneable
 	
 	public void    setBoolean(boolean bool)
 	{
-		this.number = new BigDecimal((double)(bool?1.0:0.0));
+		this.number = new BigDecimal(bool?1.0:0.0);
 	}
 	
 	public void    setInteger(long number)
